The Truncated-Uniform-Laplace (Tulap) Distribution

Distribution function and random generation for the Tulap distribution.

rtulap(n, m = 0, b = 0, q = 0)

ptulap(t, m = 0, b = 0, q = 0)

Arguments

n
number of observations

m
vector of medians

b
vector of Discrete Laplace noise parameters, obtained by \(exp(-\epsilon)\)

q
vector of truncated quantiles

t
vector of quantiles

Value

ptulap gives the distribution function and rtulap generates random derviates.

References

ptulap and rtulap are based on Awan, Jordan Alexander, and Aleksandra Slavkovic. 2020. "Differentially Private Inference for Binomial Data". Journal of Privacy and Confidentiality 10 (1). https://doi.org/10.29012/jpc.725.

Examples

rtulap(1000, 1, 0.3, 0.05)

#>    [1]  0.833913442  0.662828471  0.252100059  1.210162740  1.318766841
#>    [6]  1.826829076  0.415730145  1.177532653  0.953326463  0.552359116
#>   [11]  0.814457146  0.969858167  1.918772201  1.318651599  1.443566616
#>   [16]  3.368716549 -0.182986955  0.540157280  1.346505188  2.027305258
#>   [21] -0.493856919  0.281519568  2.973300330  1.100833983  1.269714016
#>   [26]  1.434234612  0.776589088  1.094601010 -0.157700120  1.351376971
#>   [31] -0.121231181  0.992783927  0.737911429  1.168589416 -0.515961034
#>   [36]  1.670081776  2.950699310  0.972522352  1.249202515  1.733620833
#>   [41]  1.266237268  1.307744797 -0.133311317  0.620073562  0.901452822
#>   [46] -0.848596483  2.629221468  0.816852439  2.246215231  2.611307347
#>   [51]  1.082807239  0.987019119 -0.427597187  1.031944512  0.905506434
#>   [56]  1.424987108  1.735431168  0.898741859  1.307082848 -0.724260983
#>   [61]  1.892855827  0.992169000  1.022378983  1.524550523  2.194190806
#>   [66] -0.310579606  2.281216185  0.757182817 -0.076468943  0.746079234
#>   [71] -0.029751166  0.841336169  0.635117877  1.907821877 -0.684616996
#>   [76]  0.968183135  1.472646486  0.902276946 -0.803873045  1.729971285
#>   [81] -0.900784691  0.895536173  1.643678617  2.432849072 -0.060164909
#>   [86]  0.444920169  0.876294969  0.592034719  0.929027547  0.726788016
#>   [91]  0.572085313  1.939683744  2.221751352 -0.376298726  0.516227068
#>   [96]  0.949093604  3.328397566  0.597325326  1.752724292  3.290591911
#>  [101]  1.311680153 -0.235047712  1.692723947  1.526021194  1.130683366
#>  [106]  0.933112885  1.618045400  0.861823946  0.898413260  1.365634045
#>  [111] -0.597169339  1.432438150  0.820359092  2.465180474  2.602406657
#>  [116]  0.437810779  0.820138762  1.249485268  1.767388452  0.732691002
#>  [121]  3.182651081  0.969474712  3.136511747  1.365675702  0.947501267
#>  [126]  1.355021555  0.711226542  0.277196644  1.488229194  0.596501020
#>  [131] -1.388508816 -0.379269056  2.270889420 -0.161778849  0.815656936
#>  [136]  1.435249441  2.786220934  2.252593612 -1.152665428  1.160970501
#>  [141]  2.134315362  1.169711700  1.465789858  1.557972525  1.436609981
#>  [146]  1.033475482  0.939697711  0.297762875  1.031208209  1.702220438
#>  [151]  0.611554469  1.466801522  0.888810385  1.343085690  0.101686526
#>  [156] -0.777709863  0.631106976 -0.245763212  1.319550559  1.255303041
#>  [161]  0.566844475  1.496994741  0.209319066  0.658755436  0.690822682
#>  [166]  0.849509536 -1.025187321  0.046504390  1.308845656  0.618211343
#>  [171]  0.942792427  2.000477474  2.099333799  1.451453251  1.088510900
#>  [176]  1.256516429  0.329067981  1.259293777  1.598704711  1.998436040
#>  [181]  1.397765361  0.557475951  1.332338858  1.288453310  1.255762106
#>  [186]  1.478105476  0.785280274  1.950351092  0.200265740  2.121762329
#>  [191]  1.327660083  1.430971839  3.177340604 -0.622226908  1.160213466
#>  [196]  1.225682556  1.092350019  1.484691984  2.961660586  2.631829235
#>  [201]  3.136666530  1.299758355  1.101946311  0.833645253  2.084695556
#>  [206] -0.524286638  0.098188863  1.389912168  1.473422097  1.711437221
#>  [211]  1.341834864 -0.947941415  0.040832100  1.091143052  0.307683410
#>  [216]  1.715062984  1.290182655  0.515686648  0.046209824  1.872066976
#>  [221]  2.414663497  0.980495793 -1.169870006  0.979006364  1.436799895
#>  [226] -0.886551926  0.536018071  2.399314716  0.393096951  1.562121184
#>  [231]  1.155108825  1.559146200  0.103139594  0.835791927  2.912811763
#>  [236]  0.776706976 -0.752748842  0.729947785  1.132196759  1.480120084
#>  [241]  0.968132475  0.567724703  1.150480576  0.586692624  1.081715069
#>  [246]  2.221334797  1.424051929  2.490787963  1.852580241  0.680647082
#>  [251]  0.596805891  1.164982030  0.716549680  0.640053900  1.919727572
#>  [256]  0.110485169 -0.108820956  1.368975475  0.721446523  1.084915014
#>  [261]  1.203851242  0.830455586  0.572204748  0.352055810  2.848323323
#>  [266]  0.168695818  2.339129650  0.542029931  1.124323194  2.437946605
#>  [271]  0.723237420  0.908653927  0.812696647  2.418270100  0.631626271
#>  [276]  1.456336091  1.004801215  0.602354634  0.658762728  2.048985238
#>  [281]  0.602629893  1.437647818 -0.446145883  1.076880304  1.416140727
#>  [286]  0.468825510  1.461065802  0.871676660  1.088348977  2.315559916
#>  [291] -0.533167830  1.219363465  0.489693153  0.096995187  0.452631928
#>  [296]  2.501844118  0.612414747  3.365189357  2.152328289  1.037757516
#>  [301]  1.172259003  1.134448897  0.840165631  0.766628533  2.362021839
#>  [306]  0.882699558  1.572013069 -0.048338062  0.702908403  1.020137403
#>  [311]  2.255127626  1.259825717  0.816041551 -1.335298779  1.069887567
#>  [316]  0.926426450  1.156032565  0.818476956  0.611500332  1.543995388
#>  [321]  0.529853101  0.699561803 -0.719153902  2.223452367 -0.099790877
#>  [326]  1.745243244  0.915420916  1.368552126  3.392263235  0.662290868
#>  [331]  1.274309863  1.162918539  2.109364251  1.820526443  0.780007831
#>  [336]  0.811105793  0.674836586  0.862927620  2.331343179 -1.078300044
#>  [341]  0.504649343  0.685287138  0.570574569  0.850741168  0.541448646
#>  [346]  0.088430081  0.691975894  2.021263055  0.889736891 -0.201103538
#>  [351]  1.065138289  1.189450006  1.480193708 -1.360726872  0.991436801
#>  [356]  2.480040205  0.524819632  1.238067940  1.105517661  0.463754774
#>  [361]  1.055819880 -0.032772806  1.996846364  1.895660863  0.995766778
#>  [366]  2.060915826  0.604169273 -0.592703421  1.335040702  1.721072058
#>  [371]  1.110255423  1.059552168  1.322634448  2.002069091  0.543001883
#>  [376]  0.636155165  0.785738082  1.603008385  1.470796780  1.149906439
#>  [381]  0.251979438  0.236659512 -0.352653176  1.232866235  0.680412793
#>  [386]  1.231947712  1.611477654  0.895282669  1.236835493  1.117404675
#>  [391]  2.436849608  0.785793725  3.370853838 -1.136532422  0.984836621
#>  [396]  0.668777081  0.723135890  1.235988431  0.159580651  1.953755709
#>  [401]  0.845968810  2.811818878  0.031686576 -0.485989410  0.799668808
#>  [406]  2.871143346  1.462723484  0.890179061  1.132236651 -0.450578036
#>  [411] -0.466316408  1.264262759  1.593726695  0.085938329  0.052798324
#>  [416] -0.463990910  0.447544021  0.544378269  0.895312021  0.822292117
#>  [421]  0.850016439  2.234711685  2.028120559 -0.386204272  1.357629486
#>  [426]  0.739492695  1.276938829 -0.253606271  2.683245808  0.858212552
#>  [431]  1.073974815  1.199627050 -0.122286776 -0.142700102  0.943272812
#>  [436]  2.622653021  1.051985131  0.878668121  1.027108346  0.850203981
#>  [441]  0.867578257  1.171835779  0.110398319 -0.904557048  0.712102368
#>  [446]  0.543567085  1.373910769  0.484849992  1.112528331  0.096697002
#>  [451]  3.092279169  2.208683504  0.876378844  0.103487311  2.260798417
#>  [456]  1.449172427 -1.170736478  2.104574212  1.337576153  2.334839341
#>  [461]  0.752992217  1.356589431  0.703418957  0.193791604 -1.106626629
#>  [466]  1.503252332 -0.457725590  0.662219584 -0.414859832  0.069268316
#>  [471]  1.635985471  0.616524603  1.928200383  0.776943752  0.993180373
#>  [476] -0.962738225  0.047693197  2.470997171  1.484563989  3.356675511
#>  [481]  0.577894800  2.115895833  1.508939069  2.053770138 -0.311097777
#>  [486]  3.003387762 -0.117022570  1.018204129  1.502910268  1.494398738
#>  [491]  0.823034985  0.586904650 -0.778114753  2.226891696  0.943501188
#>  [496]  1.432552925  0.633349168  1.077532802  1.304471314  1.251222088
#>  [501] -0.598317314  0.554813375  0.392915270  1.108832296  1.025948841
#>  [506]  0.996532239  1.120150539  1.487583193  0.251907112  0.868099996
#>  [511]  1.013175072  0.541353902  1.213994353  2.040577477  0.541213616
#>  [516]  1.343180893  3.130667104  0.617887965  0.660165463  1.457752543
#>  [521]  2.154524424  1.024890800  1.609052548  0.449615542  1.335532867
#>  [526]  1.289174661  3.139652105  1.475272629 -0.207572440  0.843227519
#>  [531]  0.560293627  0.958863599  1.198767520  1.040401858  2.636630645
#>  [536]  0.625681294  2.110985219  1.101067870  0.952419308  1.797864830
#>  [541]  1.040812022  0.413096565  0.659376208  2.164269446  2.390599506
#>  [546]  0.916603851  1.430194560  1.129375028  1.471249288 -0.028752474
#>  [551]  0.505193900  0.550783999  1.490471058  1.201842061  0.655015194
#>  [556]  1.499392150  0.869332432  0.595452571  0.659317364  1.247421260
#>  [561]  0.654271221  0.839769259  1.234773422  0.139795579 -0.097552212
#>  [566]  2.468271472  1.229520014  1.721175706  1.027589432  1.193842973
#>  [571]  1.140356109  1.188290846  1.390120383  0.649689997  1.385897505
#>  [576] -0.674780174  3.270286938  1.246888711 -0.425343443  2.258541018
#>  [581]  0.179130032  0.725571641  2.128512264  2.577257350  1.409402811
#>  [586]  0.651681809  0.946211150  1.034699043  1.661555656  1.109898970
#>  [591]  1.392883652  1.298529715  0.677339288  3.257388292  0.950919261
#>  [596]  0.865843280  1.212726509  1.831980056  1.093680345  0.834289856
#>  [601]  0.843762517 -0.024745831  1.740262382  0.002663674  3.364118634
#>  [606]  1.314729308  2.873111383  0.583998644  2.540823563  1.195033477
#>  [611]  1.292461826  0.660382125  1.392417502  1.253537097 -0.898167853
#>  [616]  2.625501333  1.265918018  0.962796021  0.826184288  2.209921557
#>  [621]  1.846929289  0.588086150  1.365356911  1.385096957 -0.194259620
#>  [626]  0.811153572  0.350157801  1.015642581  1.976365087  0.666315262
#>  [631]  0.339401771  0.421699834 -0.333104757  1.386087399  0.514173501
#>  [636]  0.286373856  0.006956162  1.879244419  0.713009536 -0.606419373
#>  [641]  1.319555272  0.979846911  1.106522450  1.179075758 -0.173701443
#>  [646]  0.836193201  1.174194748  2.491579492  1.173188541 -1.026815290
#>  [651]  1.190425361  1.022163972  1.034255298  1.992255930  0.170952858
#>  [656]  0.887131859  0.984868844  1.798584873  0.858386497  0.772791115
#>  [661] -0.097013373  0.686988251  1.260718023  1.218356905  0.101093522
#>  [666]  0.704381274  2.359234229  2.793248548  1.431398658  0.656935513
#>  [671]  1.469267013 -0.961461650  1.370264038  2.774797900  0.502803131
#>  [676]  1.426714036  1.620207558 -1.153029919  0.635005433  1.265513838
#>  [681] -0.014086940  1.087580371  1.996869677  0.768011284  1.279137582
#>  [686] -1.188317306  1.308527102  1.850344566  0.584839767  1.027240366
#>  [691]  0.654354660  0.513354932  0.704570009  0.959709949  1.197552218
#>  [696] -0.737393079 -0.485840846  1.444944375  1.042595792 -1.378500156
#>  [701]  0.505369025  1.220077699  0.873818595  2.313921680  0.675493301
#>  [706] -0.114877274  0.990692973  0.822206648  0.836505814  1.175582831
#>  [711]  0.735913449  0.293057269  0.882966736  0.876607339  0.541271373
#>  [716] -0.623014446  0.786595127  1.331512623  0.679440100  0.809789758
#>  [721]  1.126366857  0.605526521  2.498756090  2.255837609  1.303646139
#>  [726] -0.045497643  1.367516511  1.263282096  0.346463653  2.192007025
#>  [731]  1.419675819  1.124047702 -0.055728977  3.200336642  1.674289564
#>  [736]  2.706014640  1.313950881  0.608789870  2.483415021  1.342334251
#>  [741]  1.605557502  0.952146106  0.790993371  1.161857809  0.792693392
#>  [746]  0.919530333  1.343948454  1.304705567  1.308143277  1.027568807
#>  [751]  1.483821063  1.050372525 -0.233506432 -0.435621883  0.072203271
#>  [756]  1.101515218 -0.333937813  3.084458705  1.279350039  1.003799722
#>  [761]  0.876664010  2.515591005  2.447471513  0.066152373  1.011942199
#>  [766]  1.418848188  0.507577757  0.592567064  2.372809768  0.684952752
#>  [771]  0.510127605  2.059626922  0.664362159  0.972244077  0.855034229
#>  [776]  0.846780569  0.941592207  0.667540093  1.447108033  0.651728668
#>  [781] -1.207294030 -0.192263444  0.512480568 -1.066565225  0.830493181
#>  [786]  0.851064853  1.016603771  0.640594860 -0.357248764  1.288257637
#>  [791]  0.937601551 -0.021392884  1.873590250 -1.283648575  2.891483373
#>  [796]  0.843169858  0.568975384  0.903897540  1.552041080  2.945477036
#>  [801]  0.549780141 -0.734838637  1.378950172  0.734146304  0.997509555
#>  [806]  2.087639644  1.403996843  0.930246589 -0.364435956 -0.024649630
#>  [811]  1.361282422  0.796708409  0.451628801  2.190836894  1.472282834
#>  [816]  0.572358174  1.302932956  0.046258484  1.445903605  0.908923592
#>  [821]  1.434725216 -0.038149450  0.735632753  0.061294147  3.214823496
#>  [826]  1.169677267  0.697028611  0.805628584  2.472252346  1.643435643
#>  [831]  0.965537716  1.354474186  2.571155772  0.388044652  2.270305234
#>  [836]  1.227780907  1.704734521  1.329983958 -0.079928491  0.378590474
#>  [841]  1.282512161  1.166891018  0.773444808  2.531184097  0.469472812
#>  [846]  1.082934457  1.624533306  1.240506593  0.800179190  1.160040893
#>  [851]  1.378790433  0.770912520  0.851282692 -0.177343590  1.798284312
#>  [856]  1.191934361  1.184415556  1.486807521 -0.857130703  0.947585273
#>  [861]  2.678766401  2.413725779  1.228428771 -0.475892645 -0.375420899
#>  [866]  0.905579521  0.581521689  2.880611204  0.727881030  0.412877537
#>  [871]  1.139162693  0.574544265  2.206233451  2.411860469  1.748982221
#>  [876]  1.535838889 -0.332536535  0.104740883  1.370474275  0.401281681
#>  [881]  2.028780604  0.585626498  1.322782659  2.483850404  0.961162443
#>  [886] -0.103837638  1.469572235  0.609472246  0.979559064  0.856376159
#>  [891]  1.085202594 -0.091434447  0.157291027  1.236580148  1.687752607
#>  [896]  0.660154258 -0.080600802  0.143672850  2.282183960  1.301327362
#>  [901]  1.109082796  1.464667102 -0.077323821  0.982796400  0.546617302
#>  [906]  0.910181536  0.545569885  0.939147921 -0.161313074  0.538530515
#>  [911]  2.373988505  1.379998875 -0.422314946  0.409626871  1.279982820
#>  [916]  1.303567297  0.821321164 -0.069906952  0.639161416  2.170832755
#>  [921]  0.310407106  1.981057022  1.210416921  0.780645555  1.353523557
#>  [926]  1.435329727 -1.202272797 -0.703053436  0.845794167  0.647505625
#>  [931]  1.756182006 -0.586451363  1.236075321  1.454069805  0.448087485
#>  [936]  1.446541461  1.646166787  2.006922367  1.237385463 -1.302052981
#>  [941]  0.959800811  1.838160019  1.444574051 -0.287788257  2.261215868
#>  [946]  1.283421278  1.044175016  1.759505112  0.621113841  2.355119694
#>  [951]  2.369583778  0.947358294 -0.204276890  1.400707520  0.485904474
#>  [956]  0.732385512  1.501316256  1.302596534 -0.070895095  3.333101667
#>  [961]  0.812102548  0.911687270  1.271532991  0.453088604  1.424498814
#>  [966] -0.478962322  3.002681313  0.669246080  0.942403867  0.451275861
#>  [971]  0.233529836  0.745608970  0.811623001  1.787027895 -0.122354113
#>  [976]  0.981353665 -1.124631325  1.491971432  1.490396112  0.532765905
#>  [981]  0.572284885  0.138795480  1.266285589  3.350070267  1.343476785
#>  [986]  1.884444745  1.054221154  0.516644568  1.018226069  3.147782251
#>  [991]  1.137886690  1.050030334  1.556106715 -0.751566106  0.515016982
#>  [996]  0.089288468  2.358378678  1.411622474  1.544725962  0.629629573

ptulap(0.5, 0, 0.3, 0.05)

#> [1] 0.7834008